Fix ExprSpec cross-compilation
