+++ modules/core/src/test/scala/me/mnedokushev/zio/apache/parquet/core/filter/ExprSpec.scala
@@ -3,26 +3,12 @@ package me.mnedokushev.zio.apache.parquet.core.filter
 import me.mnedokushev.zio.apache.parquet.core.Value
 import org.apache.parquet.filter2.predicate.FilterApi
 import zio._
-import zio.schema._
-import zio.test._
 import zio.test.Assertion._
+import zio.test._
 
 import scala.jdk.CollectionConverters._
 
 object ExprSpec extends ZIOSpecDefault {
-
-  case class MyRecord(a: String, b: Int)
-  object MyRecord {
-    // TODO: figure out how to avoid specifying the type explicitly in scala 2.13 and 3
-    type A = "a"
-    type B = "b"
-    implicit val recordSchema: Schema.CaseClass2.WithFields[A, B, String, Int, MyRecord] =
-      DeriveSchema.gen[MyRecord]
-    // TODO: scala 2.12 works well without type annotation
-    // implicit val recordSchema = DeriveSchema.gen[MyRecord]
-
-    implicit val typeTag: TypeTag[MyRecord] = Derive.derive[TypeTag, MyRecord](TypeTagDeriver.default)
-  }
 
   override def spec: Spec[TestEnvironment with Scope, Any] =
     suite("ExprSpec")(
